refactor: reorganize DTOs into `projection` and `result` sub-packages
