+++ src/main/java/roomescape/controller/AdminStoreReservationController.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.*;
 import roomescape.auth.LoginMember;
 import roomescape.auth.Role;
 import roomescape.domain.Member;
-import roomescape.dto.ReservationResult;
-import roomescape.dto.StoreReservationResult;
+import roomescape.dto.result.ReservationResult;
+import roomescape.dto.result.StoreReservationResult;
 import roomescape.dto.request.ReservationUpdateRequest;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.dto.response.StoreReservationResponse;
