+++ src/main/java/roomescape/controller/ReservationController.java
@@ -4,8 +4,8 @@ import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import roomescape.auth.LoginMember;
-import roomescape.dto.ReservationResult;
-import roomescape.dto.WaitingResponseResult;
+import roomescape.dto.result.ReservationResult;
+import roomescape.dto.result.WaitingResponseResult;
 import roomescape.dto.request.ReservationCreateRequest;
 import roomescape.dto.request.ReservationUpdateRequest;
 import roomescape.dto.response.MyReservationsAndWaitsResponse;
