+++ src/main/java/roomescape/dto/result/CreatedWaitResult.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.result;
 
 import roomescape.domain.ReservationWait;
