+++ src/main/java/roomescape/dto/result/ReservationResult.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.result;
 
 import roomescape.domain.Reservation;
 import roomescape.domain.Store;
