+++ src/main/java/roomescape/service/ReservationService.java
@@ -11,10 +11,10 @@ import roomescape.domain.Member;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservationWait;
-import roomescape.dto.CreatedWaitResult;
-import roomescape.dto.ReservationResult;
-import roomescape.dto.StoreReservationResult;
-import roomescape.dto.WaitingResponseResult;
+import roomescape.dto.result.CreatedWaitResult;
+import roomescape.dto.result.ReservationResult;
+import roomescape.dto.result.StoreReservationResult;
+import roomescape.dto.result.WaitingResponseResult;
 import roomescape.exception.reservation.*;
 import roomescape.exception.reservationtime.ReservationTimeNotFoundException;
 import roomescape.exception.reservationwait.PastReservationWaitNotAllowedException;
